% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_query_intersections.R
\name{bind_cols_prefix}
\alias{bind_cols_prefix}
\title{Bind columns with a specified prefix into a single data frame}
\usage{
bind_cols_prefix(data, prefix = "seq_")
}
\arguments{
\item{data}{The input data frame.}

\item{prefix}{A character string representing the prefix of the columns to be bound.
Default is "seq_".}
}
\value{
A data frame consisting of the columns with the specified prefix bound together.
}
\description{
Given a data frame, this function binds together all columns with a specified
prefix, while maintaining their order, and returns the resulting data frame.
}
\examples{
data <- data.frame(seq_1_link_id = 1, seq_1_streetName = "A",
                   seq_2_link_id = 2, seq_2_streetName = "B",
                   count = 5)
bind_cols_prefix(data)

}
