% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_checks.R
\name{check_links_download}
\alias{check_links_download}
\title{Automatically check which Replica links have volumes associated with them.}
\usage{
check_links_download(location, folder)
}
\arguments{
\item{location}{character string pointing to top level location where data acquired from Google was saved to.}

\item{folder}{character string of name where data was automatically saved to from google data download.}
}
\value{
a returned object detailing which network links were acquired but do not have volumes associated with them.
}
\description{
This function compares the network links found in the raw network link volume table with those in the acquired Repllica network.

This function is ran automatically at the end of the query function but can be ran manually as well.
}
\examples{

#none
}
