% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_createAggByLinkSumSubsetTable.R
\name{createAggByLinkSumSubsetTable}
\alias{createAggByLinkSumSubsetTable}
\title{Aggregate and summarize link data based on count.}
\usage{
createAggByLinkSumSubsetTable(customer_name, table_agg_by_link_subset)
}
\arguments{
\item{customer_name}{The name of the BigQuery customer project.}

\item{table_agg_by_link_subset}{The name of the table containing link data.}
}
\value{
A BigQuery result containing the count of links in different categories.
}
\description{
This function generates a BigQuery query to aggregate and summarize link data
based on the count of records for each link. It categorizes links into count
ranges and provides the count for each category.
}
\examples{
queryAggByLinkSumSubset("your_project_name", "your_agg_by_link_subset_table")

}
