% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_createAggNetworkLinksTable.R
\name{createAggNetworkLinksTable}
\alias{createAggNetworkLinksTable}
\title{Aggregate network links by mode, vehicle type, and origin/destination classification.}
\usage{
createAggNetworkLinksTable(customer_name, table_trips_thru_zone, table_network)
}
\arguments{
\item{customer_name}{The name of the BigQuery customer project.}

\item{table_trips_thru_zone}{The name of the temporary table containing trips through a zone.}

\item{table_network}{The name of the temporary network table.}
}
\value{
A BigQuery result containing aggregated network links.
}
\description{
This function generates a BigQuery query to aggregate network links based on
mode, vehicle type, and the origin and destination classification of trips.
}
\examples{
queryAggNetworkLinks("your_project_name", "your_trips_thru_zone_table", "your_network_table")

}
