% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_createStudyAreaSubset.R
\name{createStudyAreaSubset}
\alias{createStudyAreaSubset}
\title{Create Study Area Subset with Spatial Index}
\usage{
createStudyAreaSubset(customer_name, wkt_geometry)
}
\arguments{
\item{wkt_geometry}{The Well-Known Text (WKT) geometry for the study area.}
}
\value{
A BigQuery result containing the study area subset.
}
\description{
This function creates a study area subset using a spatial index. It queries a BigQuery table to select records where the geometry intersects with the provided WKT geometry.
}
\examples{
wkt_geometry <- "POLYGON((x1 y1, x2 y2, x3 y3, x4 y4, x1 y1))"
study_area_subset <- create_study_area_subset(wkt_geometry)

}
