% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_createTipsByLinkIndexProcessed.R
\name{createTipsByLinkIndexProcessed}
\alias{createTipsByLinkIndexProcessed}
\title{Process Trips Table by Link Index}
\usage{
createTipsByLinkIndexProcessed(customer_name, table_trips_that_use_links)
}
\arguments{
\item{customer_name}{A character string specifying the BigQuery project name.}

\item{table_trips_that_use_links}{A character string specifying the BigQuery table containing trips with associated network links.}
}
\value{
A BigQuery result object with additional attributes representing the order and count of network links used in each trip.
}
\description{
This function processes a trips table by adding attributes that define the order in which network links are used in each trip.
}
\examples{
\dontrun{
# Example usage
result <- createTipsByLinkIndexProcessed("your_project_name", "your_trip_table_with_links")
}

}
