% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_gis_processing.R
\name{get_tigris_polys_from_replica_index}
\alias{get_tigris_polys_from_replica_index}
\title{Get polygons using Replica study area poly index.}
\usage{
get_tigris_polys_from_replica_index(
  location,
  folder,
  auto_save = F,
  network_object = NULL,
  states,
  year = 2010
)
}
\arguments{
\item{location}{character string pointing to top level location where data acquired from google was saved to.}

\item{folder}{character string of name where data was automatically saved to from google data download.}

\item{auto_save}{boolean (T/F - default F) indicating if you want the GIS layer to be saved. Default just creates an object without saving.}

\item{network_object}{network object containing links. Default is NULL or input left empty - function will use location and folder inputs to load object and then convert.}

\item{states}{character string or vector character string of state two letter abbreviations indicating which states to get polys for.}

\item{year}{integer indicating which year census polys should be acquired. Replica uses 2010 and this functions default is 2010, you should not change this.}
}
\value{
a spatial object containing polygons for study area polygons acquired from Replica.
}
\description{
Get polygons using Replica study area poly index.
}
\examples{
#none
}
