% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layer_extent_loadUnionWkt.R
\name{layer_extent_loadUnionWkt}
\alias{layer_extent_loadUnionWkt}
\title{Convert Layer to Well-Known Text (WKT) Format}
\usage{
layer_extent_loadUnionWkt(layer, layer_name)
}
\arguments{
\item{layer}{The layer to be converted to WKT format. It can be either a character (file location) or a spatial data frame.}

\item{layer_name}{A descriptive name for the layer, used for logging messages.}
}
\value{
The layer converted to WKT format with a CRS of EPSG:4326.
}
\description{
This function converts a given layer, which can be a file location (character) or a spatial data frame, to Well-Known Text (WKT) format.
It also ensures that the layer's Coordinate Reference System (CRS) is set to EPSG:4326.
}
\examples{
# Using a file location
layer <- "path/to/your_layer.shp"  # Replace with the actual file location
layer_name <- "Network Layer"
converted_layer <- convert_to_wkt(layer, layer_name)

# Using a spatial data frame
layer <- your_spatial_data_frame
layer_name <- "Study Area Layer"
converted_layer <- convert_to_wkt(layer, layer_name)

}
