% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_gis_processing.R
\name{make_agg_network_shapefile_links}
\alias{make_agg_network_shapefile_links}
\title{Convert tabular network link count objects to SF polyline objects.}
\usage{
make_agg_network_shapefile_links(
  location,
  folder,
  auto_save = F,
  aggregate_object = NULL,
  network_link_object = NULL
)
}
\arguments{
\item{location}{character string pointing to top level location where data acquired from google was saved to.}

\item{folder}{character string of name where data was automatically saved to from google data download.}

\item{auto_save}{boolean (T/F - default F) indicating if you want the GIS layer to be saved. Default just creates an object without saving.}

\item{aggregate_object}{list object containing the network link counts aggregated at different levels.}

\item{network_link_object}{network object containing links. Default is NULL or input left empty - function will use location and folder inputs to load object and then convert.}
}
\value{
a data frame and/or saved RDS file
}
\description{
This functions takes the three tabular network link counts data frames and converts them to individual SF objects contained within a list.
It can be supplied with either the list object directly or supplied location and folder strings to point to the folder the list is in.
}
\examples{
data('aggregated_network_links')
data('replica_queried_network_links')

make_agg_network_shapefile_links(
  aggregate_object = aggregated_network_links
  ,network_link_object = replica_queried_network_links
  ,auto_save = F
)
}
