% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_gis_processing.R
\name{make_network_link_layer}
\alias{make_network_link_layer}
\title{Convert tabular network file to gis layer.}
\usage{
make_network_link_layer(location, folder, auto_save = F, network_object = NULL)
}
\arguments{
\item{location}{character string pointing to top level location where data acquired from google was saved to.}

\item{folder}{character string of name where data was automatically saved to from google data download.}

\item{auto_save}{boolean (T/F - default F) inicating if you want the GIS layer to be saved. Default just creates an object without saving.}

\item{network_object}{tabular link data. Default is NULL or input left empty - function will use location and folder inputs to load object and then convert.}
}
\value{
GIS layer of network with CRS 4326
}
\description{
The network acquired from google is saved as a shapefile.
This function helps converts file from CSV to GIS layer and provides an option to save GIS object as a ".gpkg".
}
\examples{
data('replica_queried_network')

head(replica_queried_network)

make_network_link_layer(
  network_object = replica_queried_network
  ,auto_save = F
)
}
