% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_vis_objects.R
\name{make_network_map_anlto}
\alias{make_network_map_anlto}
\title{Create interactive HTML widget of network links (volumes agg. by origin and veh. type).}
\usage{
make_network_map_anlto(
  network_cntrd_object,
  poi_list = NULL,
  origin_polys = NULL
)
}
\arguments{
\item{network_cntrd_object}{list object containing aggregated link centroid networks.}

\item{poi_list}{data frame object detailing id and group attributes for polygons that network links will be displayed for. These polygons are considered Points-of-Interests and should be within the study area polygon.}

\item{origin_polys}{processed acquired_sa_polys object created for this study.}
}
\value{
an HTML widget map of network volumes
}
\description{
This function creates an interactive HTML widget which displays network volumes aggregated by trips' origin poly and vehicle type.
This HTML contains filtering and mapping features that can be used to investigate the data.
}
\examples{

data("acquired_sa_polys")
data("poi_list")
data("network_centroid_aggregation_list")

make_network_map_anlto(
  network_cntrd_object = network_centroid_aggregation_list
  ,poi_list = poi_list
  ,origin_polys = acquired_sa_polys
)
}
