% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_query_intersections.R
\name{make_space_2}
\alias{make_space_2}
\title{Create a string of repeated characters for indentation purposes}
\usage{
make_space_2(with = "+", n = 50, c = "", last = T)
}
\arguments{
\item{with}{A character to repeat.}

\item{n}{An integer specifying the number of times to repeat the character.}

\item{c}{A string to separate each repeated character. Defaults to an empty string.}

\item{last}{A boolean indicating whether the resulting string should end with a newline character. Defaults to TRUE.}
}
\value{
A string of repeated characters separated by the specified separator string.
}
\description{
This function creates a string of a specified length containing a specified character. It is commonly used to create indentation in printed output.
}
\examples{
make_space_2(n = 5, with = " ")
make_space_2(n = 5, with = "-", c = "", last = FALSE)

}
