% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_gis_processing.R
\name{make_trip_origin_point_layer}
\alias{make_trip_origin_point_layer}
\title{Convert tabular first link file to gis layer.}
\usage{
make_trip_origin_point_layer(
  location,
  folder,
  auto_save = F,
  first_links_object = NULL
)
}
\arguments{
\item{location}{character string pointing to top level location where data acquired from google was saved to.}

\item{folder}{character string of name where data was automatically saved to from Google data download.}

\item{auto_save}{boolean (T/F - default F) indicating if you want the GIS layer to be saved. Default just creates an object without saving.}

\item{network_object}{network object containing links. Default is NULL or input left empty - function will use location and folder inputs to load object and then convert.}
}
\value{
GIS layer of network with CRS 4326
}
\description{
The network acquired from google is saved as a shapefile.
This function helps converts file from CSV to GIS layer and provides an option to save GIS object as a ".gpkg".
}
\examples{
data('replica_queried_network_links')

head(replica_queried_network_links)

make_network_centroid_layer(
  location = "data/req_dev"
  ,folder = "data_20230106_154119"
  ,network_object = replica_queried_network_links
  ,auto_save = F
)
}
