% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_replica_movements.R
\name{map_replica_movements}
\alias{map_replica_movements}
\title{Maps discrete movement patterns from Replica}
\usage{
map_replica_movements(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
A map object created using mapview package
}
\description{
This function takes a data frame and creates a map using the mapview package.
The map shows the distribution of the 'order' column values, grouped by the 'flag_grp' column values.
}
\examples{
map_replica_movements(data)

}
