% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_query_intersections.R
\name{merge_cols}
\alias{merge_cols}
\title{Merge Columns and Create Movement Descriptions}
\usage{
merge_cols(data)
}
\arguments{
\item{data}{A data frame to be processed.}
}
\value{
A processed data frame with additional "mvmnt_desc" and "mvmnt_desc_fl"
columns.
}
\description{
This function takes a data frame as input and merges all columns whose names
start with "seq_" with a separate table called "table_network_data_simp".
For each merged column, it also creates a new column with a concatenated
string of the corresponding "streetName" values from the "table_network_data_simp"
table. Finally, the function creates two new columns in the input data frame:
"mvmnt_desc" and "mvmnt_desc_fl". The "mvmnt_desc" column contains a concatenated
string of all street names for each row, and the "mvmnt_desc_fl" column
contains a concatenated string of only the first and last street names for each row.
}
\examples{
my_data <- data.frame(seq_1 = c(1,2,3), seq_2 = c(4,5,6), seq_3 = c(7,8,9),
                      stableEdgeId = c(1,2,3), streetName = c("First Street",
                      "Second Street", "Third Street"))
merge_cols(my_data)

}
