% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_query_intersections.R
\name{order_map}
\alias{order_map}
\title{Maps discrete movement patterns from Replica}
\usage{
order_map(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
A map object created using mapview package
}
\description{
This function takes a data frame and creates a map using the mapview package.
The map shows the distribution of the 'order' column values, grouped by the 'flag_grp' column values.
}
\examples{
order_map(data)

}
