% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_query_intersections.R
\name{prompt_jitter_factor}
\alias{prompt_jitter_factor}
\title{Prompt User for Jitter Factor}
\usage{
prompt_jitter_factor()
}
\value{
A numeric value indicating the jitter factor entered by the user.
}
\description{
This function prompts the user to enter a jitter factor value between 0.00001
and 0.01. The function accepts the user's input and handles invalid input by
prompting the user again.
}
\examples{
prompt_jitter_factor()

}
\keyword{factor,}
\keyword{input}
\keyword{jitter}
\keyword{prompt,}
\keyword{user}
