% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_query_intersections.R
\name{query_replica_mvmnt_patterns}
\alias{query_replica_mvmnt_patterns}
\title{Query movement patterns for a replica network}
\usage{
query_replica_mvmnt_patterns(
  network_table,
  trip_table,
  mode_type,
  customer_name,
  jitter_factor
)
}
\arguments{
\item{network_table}{A character string specifying the name of the table containing network data}

\item{trip_table}{A character string specifying the name of the table containing trip data}

\item{mode_type}{A character string specifying the type of mode to query}

\item{customer_name}{A character string specifying the name of the customer}

\item{jitter_factor}{A numeric value specifying the amount of jitter to apply to the network display}
}
\value{
A processed data frame of movement patterns
}
\description{
This function queries movement patterns  replica network using a specified \code{network_table},
\code{trip_table}, \code{mode_type}, \code{customer_name}, and \code{jitter_factor}. It uses
the \code{view_replica_network} function to retrieve the network data, applies a jitter factor to
display the network, prompts the user to select links for each movement, performs a table query, and
performs quality control and processing operations.
}
\examples{
query_replica_mvmnt_patterns(network_table = "network_data", trip_table = "trip_data",
                             mode_type = "PRIVATE_AUTO", customer_name = "ACME",
                             jitter_factor = 0.05)

}
