% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_trips_by_OD_polygons.R
\name{query_trips_by_OD_polygons}
\alias{query_trips_by_OD_polygons}
\title{Query Replica trips by origin or destination polygons}
\usage{
query_trips_by_OD_polygons(
  network_table,
  trip_table,
  customer_name,
  query_links = c("highway", "corridor", "road", "motorway", "motorway_link", "trunk",
    "trunk_link", "primary", "primary_link", "secondary", "secondary_link", "tertiary",
    "tertiary_link"),
  mode_type = c("PASSENGER_CAR", "PRIVATE_AUTO", "COMMERCIAL", "CARPOOL", "WALKING",
    "BIKING", "PUBLIC_TRANSIT", "ON_DEMAND_AUTO", "OTHER_TRAVEL_MODE"),
  index_od,
  study_area,
  poly_gen_att
)
}
\arguments{
\item{network_table}{A string specifying the network table from the API to use.}

\item{trip_table}{A string indicating the trip table to use.}

\item{customer_name}{A string representing the customer name.}

\item{query_links}{A character vector specifying types of links to query (e.g., 'highway', 'corridor', 'road', 'motorway', 'motorway_link', 'trunk', 'primary', 'primary_link', 'secondary', 'secondary_link').}

\item{mode_type}{A character vector containing the mode types to filter the data (e.g., 'COMMERCIAL', 'PRIVATE_AUTO', 'ON_DEMAND_AUTO').}

\item{index_od}{A vector of two strings indicating locational attributes to query with (e.g., c("origin_long", "origin_lat")).}

\item{study_area}{A spatial polygon denoting the extent of the study area - this will be used to query network links, Should be as small as possible to limit the number of links the user downloads.}

\item{poly_gen_att}{A spatial data frame (using the SF package) that denotes polygons to be used as origin or destination locations for the query.}
}
\value{
A list containing the results of the Replica Mobility Data Platform query. The list consists of two elements:

 1. `table_od_poly`: A BigQuery table string representing the table of OD polygons.
 2. `table_trips_links`: A BigQuery table string representing the table of trip links.
}
\description{
Query Replica trips by origin or destination polygons
}
\details{
The primary focus of this function is to retrieve trips that either originate or terminate within specific polygons, which can be custom polygons or block groups, etc.

It is ideal for capturing trips that have their starting or ending points within these defined polygons.
If you need to query trips that pass through a specific zone but don't necessarily start or end in that zone, please refer to the separate function designed for that purpose.

This function facilitates the connection to the Replica Mobility Data Platform API, allowing users to query and retrieve transportation data. It requires several inputs for customization, see below.
}
\examples{
\dontrun{
# Example usage:
result <- query_trips_by_OD_polygons(
  network_table = "replica-customer.northwest.northwest_2019_Q4_network_segments",
  trip_table = "replica-customer.northwest.northwest_2019_Q4_thursday_trip",
  customer_name = "replica-customer",
  mode_type = c('COMMERCIAL'),
  query_links = c("highway", "corridor", "road", "motorway", "motorway_link", "trunk",
                  "primary", "primary_link", "secondary", "secondary_link"),
  index_od = c("origin_long", "origin_lat"),
  study_area = study_area_sf,
  poly_gen_att = poly_gen_att_sf
)
}

}
\seealso{
More information about the Replica Mobility Data Platform can be found at [Replica Mobility Data Platform](https://www.replicamobility.com/).
}
