% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_google_queries.R,
%   R/replica_temp_tbl_name.R
\name{replica_temp_tbl_name}
\alias{replica_temp_tbl_name}
\title{Extract project name, dataset, and tablename from google big query object.}
\usage{
replica_temp_tbl_name(object)

replica_temp_tbl_name(object)
}
\arguments{
\item{object}{}
}
\value{
a character string

a character string
}
\description{
When BQuery objects (tables) are made using bq_project_query() they are given temporary dataset and table names.
This function extracts all of a table's identifying elements and glues them into a singular string.
This string can be used in subsequent queires.

When BQuery objects (tables) are made using bq_project_query() they are given temporary dataset and table names.
This function extracts all of a table's identifying elements and glues them into a singular string.
This string can be used in subsequent queires.
}
\details{
Function is helpful when making mulitstepped queries where temporary tables are created and subsequently interacted with.

Function is helpful when making mulitstepped queries where temporary tables are created and subsequently interacted with.
}
\examples{

temp = list(username = "replica_fake_name"
,dataset = "replica_temp_dataset"
,table = "replica_temp_table")

replica_temp_tbl_name(temp)

temp = list(username = "replica_fake_name"
,dataset = "replica_temp_dataset"
,table = "replica_temp_table")

replica_temp_tbl_name(temp)
}
