% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replica_test_connection.R
\name{replica_test_connection}
\alias{replica_test_connection}
\title{Test Connection to Replica Tables Before Running a Larger Query}
\usage{
replica_test_connection(
  network_table = network_table,
  trip_table = trip_table,
  customer_name = customer_name
)
}
\arguments{
\item{network_table}{Name of the Replica network table.}

\item{trip_table}{Name of the Replica trip table.}

\item{customer_name}{Name of the customer owning the Replica tables.}
}
\value{
Invisible. Throws an error if the connection to one or more tables fails.
}
\description{
This function tests the connection to the specified Replica network and trip tables
by attempting to download a row from each table. It ensures that the tables exist
and can be accessed before running a more substantial query. If the tables do not
exist or if there is an error during the connection attempt, an error message is provided.
}
\details{
The function uses the \code{bigrquery} package to query the Replica network
and trip tables by attempting to download a single row from each table. It checks
the existence and accessibility of the tables before proceeding with a more extensive query.
If an error occurs during the connection attempt, an error message is displayed.
}
\examples{
# Example usage:
# replica_test_connection(network_table = "your_network_table",
#                         trip_table = "your_trip_table",
#                         customer_name = "your_customer_name")

}
