% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplc_check_and_log_queired_links.R
\name{rplc_check_and_log_queired_links}
\alias{rplc_check_and_log_queired_links}
\title{Check and log information about the number of unique links per Highway type.}
\usage{
rplc_check_and_log_queired_links(counts_object, query_links, logger_object)
}
\arguments{
\item{counts_object}{A data frame containing the count of unique links per
highway type.}

\item{query_links}{A vector of requested link types.}

\item{logger_object}{A logger_object object for logging messages.}
}
\value{
None.
}
\description{
This function performs checks on the provided data to ensure that the number
of unique links is not zero and logs information accordingly.
}
