% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplc_checkValidTableConnections.R
\name{rplc_checkValidTableConnections}
\alias{rplc_checkValidTableConnections}
\title{Function to download sample network and trip tables, perform validity checks, and notify users of errors.}
\usage{
rplc_checkValidTableConnections(
  prefix_origin,
  prefix_dest,
  network_table,
  trip_table
)
}
\arguments{
\item{prefix_origin}{The prefix for the origin location columns.}

\item{prefix_dest}{The prefix for the destination location columns.}

\item{network_table}{The network table to be downloaded.}

\item{trip_table}{The trip table to be downloaded.}
}
\value{
This function performs checks on the downloaded data, and if errors are found, it will stop the function and notify the user.
}
\description{
This function downloads a sample of both network and trip tables, puts them into temporary tables, and checks for errors in the data.
}
\details{
The function first downloads sample data from both the network and trip tables, setting the maximum number of records to 1. It then checks whether the downloaded trip table contains valid column names based on the provided prefixes for origin and destination, followed by "_lat" and "_lng".

If either of the downloaded tables has no records, the function will flag an error, stop the execution, and notify the user.

Additionally, the function checks whether the column names derived from the prefixes for origin and destination are valid call names in the temporary data frame.
}
\examples{
\dontrun{
prefix_origin <- "origin"
prefix_dest <- "destination"
network_table <- "your_network_table_name"
trip_table <- "your_trip_table_name"
download_and_check(prefix_origin, prefix_dest, network_table, trip_table)
}
}
