% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rplc_jitter_factor_prompt.R
\name{rplc_jitter_factor_prompt}
\alias{rplc_jitter_factor_prompt}
\title{Prompt User for Jitter Factor}
\usage{
rplc_jitter_factor_prompt()
}
\value{
A numeric value indicating the jitter factor entered by the user.
}
\description{
This function prompts the user to enter a jitter factor value between 0.00001
and 0.01. The function accepts the user's input and handles invalid input by
prompting the user again.
}
\examples{
\dontrun{
rplc_jitter_factor_prompt()
}
}
