% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seattleFreight_replica_network_links}
\alias{seattleFreight_replica_network_links}
\title{Replica Network Links (Polyline)}
\format{
An `sf` object with 8 variables.
\describe{
  \item{stableEdgeId}{Replica link ID.}
  \item{streetName}{Assigned street name.}
  \item{distance}{Link length in meters.}
  \item{osmid}{OSM roadway ID.}
  \item{highway}{Highway class label.}
  \item{flag_contains}{TRUE if inside network study area.}
  \item{flag_highway}{Internal tag identifying freight-relevant links.}
  \item{geometry}{LINESTRING geometry.}
}
}
\usage{
seattleFreight_replica_network_links
}
\description{
Replica Network Links (Polyline)
}
\details{
This is the polyline representation of the queried Replica network.
It contains only a subset of attributes from the raw Replica table, enough for mapping and joining.
Used for visualizing links and merging aggregated network volumes.
Created by processing the raw queried network table into sf lines.
}
\keyword{datasets}
