% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seattleFreight_replica_network_points}
\alias{seattleFreight_replica_network_points}
\title{Replica Network Points (Centroids)}
\format{
An `sf` object with 8 variables.
\describe{
  \item{stableEdgeId}{Replica link ID.}
  \item{streetName}{Assigned street name.}
  \item{osmid}{OSM roadway ID.}
  \item{highway}{Highway class.}
  \item{flag_contains}{TRUE if inside network study area.}
  \item{flag_highway}{Internal freight-highway tag.}
  \item{geometry}{POINT geometry representing link midpoint.}
}
}
\usage{
seattleFreight_replica_network_points
}
\description{
Replica Network Points (Centroids)
}
\details{
These points represent the centroid of each Replica link.
They are used because point layers work better with Crosstalk/Leaflet widgets than polylines.
Contains the same minimal attributes as the network links table but with point geometry.
Helpful for interactive filtering, dashboards, and map popups.
}
\keyword{datasets}
