% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seattleFreight_replica_network_qs}
\alias{seattleFreight_replica_network_qs}
\title{Raw Queried Replica Network Table (QS)}
\format{
A tibble with 14 variables.
\describe{
  \item{stableEdgeId}{Replica link ID.}
  \item{startLat, startLon}{Starting coordinates.}
  \item{endLat, endLon}{Ending coordinates.}
  \item{streetName}{Street name.}
  \item{distance}{Length in meters.}
  \item{osmid}{OSM ID.}
  \item{speed}{Assigned link speed.}
  \item{flags}{Allowed-mode flags.}
  \item{lanes}{Number of lanes.}
  \item{highway}{Highway class.}
  \item{geometry}{WKT LINESTRING.}
  \item{flag_contains}{TRUE if link is inside study area.}
}
}
\usage{
seattleFreight_replica_network_qs
}
\description{
Raw Queried Replica Network Table (QS)
}
\details{
This is the direct output from Replica’s network table for the user-defined study area.
It serves as the base table for building link-sf objects and extracting midpoints.
Includes geometry, roadway attributes, and link metadata needed for downstream processing.
All spatial network products originate from this dataset.
}
\keyword{datasets}
