% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_createNetworkLinkCountTable.R
\name{sql_createNetworkLinkCountTable}
\alias{sql_createNetworkLinkCountTable}
\title{Query the count of network links by highway type.}
\usage{
sql_createNetworkLinkCountTable(customer_name, table_network = table_network)
}
\arguments{
\item{customer_name}{The name of the BigQuery customer project.}

\item{table_network}{The name of the temporary table containing network data.}
}
\value{
A BigQuery result containing the count of network links by highway type.
}
\description{
This function generates a BigQuery query to calculate the count of network
links by highway type using a temporary table created previously.
}
\examples{
\dontrun{
# none

}
}
