% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_createNetworkTable.R
\name{sql_createNetworkTable}
\alias{sql_createNetworkTable}
\title{Create a temporary BigQuery table for network data.}
\usage{
sql_createNetworkTable(
  customer_name,
  network_table = network_table,
  links_pro = links_pro,
  wkt_object,
  highway_regrex = NULL
)
}
\arguments{
\item{customer_name}{The name of the BigQuery customer project.}

\item{network_table}{The name of the network data table in BigQuery.}

\item{links_pro}{A vector of link types to filter the data.}

\item{list_wkt_objects}{A list of well-known text (WKT) geometry objects
used to define the study area.}
}
\value{
A temporary BigQuery table containing filtered network data.
}
\description{
This function generates a BigQuery query to create a temporary table by
selecting network data based on a study area defined by geometry and link types.
}
\examples{
\dontrun{
sql_createNetworkTable(
"your_project_name", "your_network_table"
,c("highway", "street"), list_of_wkt_objects)

}
}
