% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_createStudyAreaSubset.R
\name{sql_createStudyAreaSubset}
\alias{sql_createStudyAreaSubset}
\title{Create Study Area Subset with Spatial Index}
\usage{
sql_createStudyAreaSubset(customer_name, wkt_geometry)
}
\arguments{
\item{wkt_geometry}{The Well-Known Text (WKT) geometry for the study area.}
}
\value{
A BigQuery result containing the study area subset.
}
\description{
This function creates a study area subset using a spatial index. It queries a BigQuery table to select records where the geometry intersects with the provided WKT geometry.
}
\examples{
\dontrun{
# none

}
}
