% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_createTripNetworkMatchTable.R
\name{sql_createTripNetworkMatchTable}
\alias{sql_createTripNetworkMatchTable}
\title{Query distinct activity IDs based on network link matching.}
\usage{
sql_createTripNetworkMatchTable(
  customer_name,
  trip_table = trip_table,
  mode_type_pro = mode_type_pro,
  table_network = table_network
)
}
\arguments{
\item{customer_name}{The name of the BigQuery customer project.}

\item{trip_table}{The name of the trip data table in BigQuery.}

\item{mode_type_pro}{A vector of mode types to filter the data.}

\item{table_network}{The name of the temporary table containing network data.}
}
\value{
A BigQuery result containing distinct activity IDs.
}
\description{
This function generates a BigQuery query to obtain distinct activity IDs
where network link IDs match those in the temporary network table.
}
\examples{
\dontrun{
# none

}
}
