% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_createTripsThruZoneTable_mvmntPat.R
\name{sql_createTripsThruZoneTable_mvmntPat}
\alias{sql_createTripsThruZoneTable_mvmntPat}
\title{Query trips through a zone and classify origin and destination.}
\usage{
sql_createTripsThruZoneTable_mvmntPat(
  customer_name,
  trip_table,
  table_sa_poly_index,
  activity_id_pro,
  table_activity_mvmnt_seq_list_comb
)
}
\arguments{
\item{customer_name}{The name of the BigQuery customer project.}

\item{trip_table}{The name of the trip data table in BigQuery.}

\item{table_sa_poly_index}{The name of the temporary table containing study area blockgroups.}

\item{mode_type_pro}{A vector of mode types to filter the data.}
}
\value{
A BigQuery result containing trips through a zone with classified origin
and destination points.
}
\description{
This function generates a BigQuery query to select trips through a zone and
classify their origin and destination points based on whether they are inside
the specified spatial index. It also categorizes the trips as internal or external.
}
\examples{
queryTripsThruZone("your_project_name", "your_trip_table", c("mode1", "mode2"))

}
