% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_vis_objects.R
\name{viz_static_ntwrk_map_anl}
\alias{viz_static_ntwrk_map_anl}
\title{Make interactive network link volume maps using MapView.}
\usage{
viz_static_ntwrk_map_anl(spatial_agg_object, return_leaflet)
}
\arguments{
\item{spatial_agg_object}{spatial links objects of aggregated network. Has to be links, centroids will print but aesthetic options in Mapview code are specific to links.}

\item{return_leaflet}{Boolean (T/F) indicating if a leaflet or Mapview object should be returned. Default F returns a Mapview object, these can be added together.}
}
\value{
a leaflet object made using MapView API depicting links
}
\description{
This is a helper function that quickly makes an interactive map of aggregated network link volumes.
All functions in the viz_static_ntwrk_map_% function family makes a singular leaflet/mapview object.

This is another mapping option if you wish to map network links and view link volumes without making an interactive, HTML elements.
}
\details{
ANLT (Aggregated Network Links).
}
\examples{

data_temp = aggregate_network_links(
agg_count_object = table_agg_by_link_subset_limited
,network_object = replica_queried_network_links
)

viz_static_ntwrk_map_anltpt(
  spatial_agg_object = data_temp
  )
}
