% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_vis_objects.R
\name{viz_static_ntwrk_map_anlto_grp}
\alias{viz_static_ntwrk_map_anlto_grp}
\title{Make interactive network link volume maps using MapView - maps ANLT network links (Aggregated Network Links by vehicle Type and Origin poly).}
\usage{
viz_static_ntwrk_map_anlto_grp(spatial_agg_object, poi_list, return_leaflet)
}
\arguments{
\item{spatial_agg_object}{spatial links objects of aggregated network. Has to be links, centroids will print but aesthetic options in Mapview code are specific to links.}

\item{poi_list}{data frame object detailing id and group attributes for polygons that network links will be displayed for. These polygons are considered Points-of-Interests and should be within the study area polygon.}

\item{return_leaflet}{Boolean (T/F) indicating if a leaflet or Mapview object should be returned. Default F returns a Mapview object, these can be added together.}
}
\value{
a leaflet object made using MapView API depicting links
}
\description{
This is a helper function that quickly makes an interactive map of aggregated network link volumes.
All functions in the viz_static_ntwrk_map_% function family makes a singular leaflet/mapview object.

This is another mapping option if you wish to map network links and view link volumes without making an interactive, HTML elements.
}
\examples{

data_temp = aggregate_network_links(
agg_count_object = table_agg_by_link_subset_limited
,network_object = replica_queried_network_links
)

viz_static_ntwrk_map_anlto_grp(
  spatial_agg_object = data_temp
  )
}
