% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_network_trips_using_bbox.R
\name{query_network_trip_using_bbox}
\alias{query_network_trip_using_bbox}
\title{Easily query and download Replica trip and network data using the extent of a map layer.}
\usage{
query_network_trip_using_bbox(
  bb_network_layer,
  bb_sa_layer,
  query_links = c("highway", "corridor", "road", "motorway", "motorway_link", "trunk",
    "trunk_link", "primary", "primary_link", "secondary", "secondary_link", "tertiary",
    "tertiary_link"),
  mode_type = c("PASSENGER_CAR", "PRIVATE_AUTO", "COMMERCIAL", "CARPOOL", "WALKING",
    "BIKING", "PUBLIC_TRANSIT", "ON_DEMAND_AUTO", "OTHER_TRAVEL_MODE"),
  data_set_location,
  data_set_period,
  data_set_day,
  customer_name,
  prefix_origin = "origin",
  prefix_dest = "destination",
  input_folder_write,
  input_folder_suffix = NULL,
  max_record = 1000,
  input_highway_regrex = NULL,
  check_aux_tables = NULL
)
}
\arguments{
\item{bb_network_layer}{a string of the relative path to the bounding box object detailing the extent of the road network. This will be used to query trips with. This should be made using sf::st_bbox().}

\item{bb_sa_layer}{a string of the relative path to the bounding box object detailing the extent of the study area. This will be used to perform aggregations. This should be made using sf::st_bbox(). This input is generally larger or identical to [bb_network_layer].}

\item{query_links}{a vector of character strings detailing which type of network links to use in the query. These will be used to depict any network link volume graphics and to filter trips. For the latter, any trip that uses these roads within the study area will be included in any reporting.}

\item{mode_type}{a vector of character strings detailing which mode to query, e.g., 'PASSENGER_CAR', 'PRIVATE_AUTO', 'COMMERCIAL', etc.}

\item{data_set_location}{a string indicating the region or location table to use, e.g., "northwest."}

\item{data_set_period}{a string in year_quarter format, e.g., "2023_Q1."}

\item{data_set_day}{a string that should be either "thursday" or "saturday."}

\item{customer_name}{a character string indicating the Google account that will be billed.}

\item{prefix_origin}{a string defining the prefix for Latin long coordinates for each record for the origin (e.g., "origin").}

\item{prefix_dest}{a string defining the prefix for Latin long coordinates for each record for the destination (e.g., "destination").}

\item{input_folder_write}{a character string indicating a directory where you want log files and data to be saved in CSV format. A new folder containing function outputs will be created there, named using the convention 'data_[sys.datetime].'}

\item{max_record}{an integer indicating the maximum number of records to be written out for each data acquired by Google. The default is 1000. Use Inf if you do not want to limit the download. This is useful if you want to check the data before downloading large amounts of data. Google tables are still made in full so they can be manually acquired without rerunning this function (see log files).}

\item{query_network}{a boolean (T/F) indicating whether the network should be downloaded at function runtime. The default is 'T,' but it can be beneficial to wait and see how many links will be downloaded before downloading the network. The network can also be downloaded later as log files record big query table IDs.}
}
\value{
This function does not return any objects in R. It creates a log file and data in the folder destination in CSV format.
}
\description{
This function takes a bounding box and Google Big Query inputs and returns data back to you.
Function currently only works for data using block groups as it is hardcoded in some of the queries. Changes can be made to the query such that other polygon shapes can be used.

This function creates a log file that records inputs, the query sent to Google, and IDs for each Google table that is created in the multi-step query. All of this information can be accessed at later points in time. Table IDs in the log file can be used for auditing purposes or copied and pasted for use in the Google Big Query console GUI.

Reprinted: 20231107
}
\examples{
\dontrun{

#none

}
}
